todoapp/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-redis-cache==3.0.0
djangorestframework==3.12.4
pkg-resources==0.0.0
pytz==2021.1
redis==3.5.3
sqlparse==0.4.2
typing-extensions==3.10.0.2